+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * nostr: change `Tag::parse` arg from `Vec<S>` to `&[S]` ([Yuki Kishimoto])
 * nostr: allow to parse public key from NIP21 uri with `PublicKey::parse` ([Yuki Kishimoto])
 * nostr: allow to parse event ID from NIP21 uri with `EventId::parse` ([Yuki Kishimoto])
+* nostr: construct `GenericTagValue` based on `SingleLetterTag` in `deserialize_generic_tags` ([Yuki Kishimoto])
 * nostr: bump `bitcoin` to `0.31` ([Yuki Kishimoto])
 * sdk: bump `lnurl-pay` to `0.4` ([Yuki Kishimoto])
 * nwc: avoid to open and close subscription for every request ([Yuki Kishimoto])
@@ -28,7 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * nostr: `Tag::content` return always `None` when `Tag::Generic` ([Yuki Kishimoto])
-* nostr: fix `GenericTagValue` deserialization ([Yuki Kishimoto])
 
 ### Removed